Consolidate training deployment selection and assembly
